requirements.txt
discord-webhook
flask
python-dotenv
art
GitPython
requests